+++ src/components/books/Books.js
@@ -1,37 +1,12 @@
+import PropTypes from 'prop-types';
 import Book from './Book';
 
-const books = [
-  {
-    category: 'Action',
-    title: 'The Hunger Games',
-    author: 'Suzanne Collins',
-    progress: {
-      currentChapter: 'Chapter 17',
-      completed: '64',
-    },
-  },
-  {
-    category: 'Science Fiction',
-    title: 'Dune',
-    author: 'Frank Herbert',
-    progress: {
-      currentChapter: 'Chapter 3: "A Lesson Learned"',
-      completed: '8',
-    },
-  },
-  {
-    category: 'Economy',
-    title: 'Capital in the Twenty-First Century',
-    author: 'Suzanne Collins',
-    progress: {
-      currentChapter: 'Introduction',
-      completed: '0',
-    },
-  },
-];
-
-const Books = () => (
+const Books = ({ books }) => (
   <ul>{books.map((book) => <li key={book.title}><Book book={book} /></li>)}</ul>
 );
+
+Books.propTypes = {
+  books: PropTypes.arrayOf(Book.propTypes.book).isRequired,
+};
 
 export default Books;
